Generalize CommandFactory and QueryFactory by introducing and abstract `ActorRequestFactory` base; migrate the static factory methods from `Queries` to `QueryFactory` to take the actor context into account.
