+++ client/src/main/java/org/spine3/client/Queries.java
@@ -27,9 +27,9 @@ import org.spine3.type.TypeName;
 import org.spine3.type.TypeUrl;
 
 import javax.annotation.Nullable;
-import java.util.Arrays;
 import java.util.Set;
 
+import static com.google.common.base.Preconditions.checkNotNull;
 import static org.spine3.client.Queries.Targets.allOf;
 import static org.spine3.client.Queries.Targets.someOf;
 
@@ -45,103 +45,18 @@ public class Queries {
         // Prevent instantiation of this utility class.
     }
 
-    /**
-     * Create a {@link Query} to read certain entity states by IDs with the {@link FieldMask}
-     * applied to each of the results.
-     *
-     * <p>Allows to specify a set of identifiers to be used during the {@code Query} processing.
-     * The processing results will contain only the entities, which IDs are present among
-     * the {@code ids}.
-     *
-     * <p>Allows to set property paths for a {@link FieldMask}, applied to each of the query
-     * results. This processing is performed according to the
-     * <a href="https://developers.google.com/protocol-buffers/docs/reference/google.protobuf#google.protobuf.FieldMask">FieldMask specs</a>.
-     *
-     * <p>In case the {@code paths} array contains entries inapplicable to the resulting entity
-     * (e.g. a {@code path} references a missing field), such invalid paths are silently ignored.
-     *
-     * @param entityClass the class of a target entity
-     * @param ids         the entity IDs of interest
-     * @param paths       the property paths for the {@code FieldMask} applied to each of results
-     * @return an instance of {@code Query} formed according to the passed parameters
-     */
-    public static Query readByIds(Class<? extends Message> entityClass,
-                                  Set<? extends Message> ids,
-                                  String... paths) {
-        final FieldMask fieldMask = FieldMask.newBuilder()
-                                             .addAllPaths(Arrays.asList(paths))
-                                             .build();
-        final Query result = composeQuery(entityClass, ids, fieldMask);
-        return result;
-    }
-
-    /**
-     * Create a {@link Query} to read all entity states with the {@link FieldMask}
-     * applied to each of the results.
-     *
-     * <p>Allows to set property paths for a {@link FieldMask}, applied to each of the query
-     * results. This processing is performed according to the
-     * <a href="https://developers.google.com/protocol-buffers/docs/reference/google.protobuf#google.protobuf.FieldMask">FieldMask specs</a>.
-     *
-     * <p>In case the {@code paths} array contains entries inapplicable to the resulting entity
-     * (e.g. a {@code path} references a missing field), such invalid paths are silently ignored.
-     *
-     * @param entityClass the class of a target entity
-     * @param paths       the property paths for the {@code FieldMask} applied to each of results
-     * @return an instance of {@code Query} formed according to the passed parameters
-     */
-    public static Query readAll(Class<? extends Message> entityClass, String... paths) {
-        final FieldMask fieldMask = FieldMask.newBuilder()
-                                             .addAllPaths(Arrays.asList(paths))
-                                             .build();
-        final Query result = composeQuery(entityClass, null, fieldMask);
-        return result;
-    }
-
-    /**
-     * Create a {@link Query} to read certain entity states by IDs.
-     *
-     * <p>Allows to specify a set of identifiers to be used during the {@code Query} processing.
-     * The processing results will contain only the entities, which IDs are present among
-     * the {@code ids}.
-     *
-     * <p>Unlike {@link Queries#readByIds(Class, Set, String...)}, the {@code Query} processing
-     * will not change the resulting entities.
-     *
-     * @param entityClass the class of a target entity
-     * @param ids         the entity IDs of interest
-     * @return an instance of {@code Query} formed according to the passed parameters
-     */
-    public static Query readByIds(Class<? extends Message> entityClass,
-                                  Set<? extends Message> ids) {
-        return composeQuery(entityClass, ids, null);
-    }
-
-    /**
-     * Create a {@link Query} to read all states of a certain entity.
-     *
-     * <p>Unlike {@link Queries#readAll(Class, String...)}, the {@code Query} processing will
-     * not change the resulting entities.
-     *
-     * @param entityClass the class of a target entity
-     * @return an instance of {@code Query} formed according to the passed parameters
-     */
-    public static Query readAll(Class<? extends Message> entityClass) {
-        return composeQuery(entityClass, null, null);
-    }
-
-    private static Query composeQuery(Class<? extends Message> entityClass,
+    static Query.Builder queryBuilderFor(Class<? extends Message> entityClass,
                                       @Nullable Set<? extends Message> ids,
                                       @Nullable FieldMask fieldMask) {
+        checkNotNull(entityClass);
+
         final Target target = ids == null ? allOf(entityClass) : someOf(entityClass, ids);
         final Query.Builder queryBuilder = Query.newBuilder()
                                                 .setTarget(target);
         if (fieldMask != null) {
             queryBuilder.setFieldMask(fieldMask);
         }
-        final Query result = queryBuilder
-                .build();
-        return result;
+        return queryBuilder;
     }
 
     /**
@@ -153,6 +68,8 @@ public class Queries {
      * @return the URL of the type of the query {@linkplain Query#getTarget() target}
      */
     public static TypeUrl typeOf(Query query) {
+        checkNotNull(query);
+
         final Target target = query.getTarget();
         final TypeName typeName = TypeName.of(target.getType());
         final TypeUrl type = typeName.toUrl();
